Чиним RSA

+++ src/libraly/data/ElGamaliaHashData.kt
@@ -1,3 +1,3 @@
 package libraly.data
 
-class ElGamaliaHashData(override val m: Byte, val r: Double, val s: Double) : HashDataRule
+class ElGamaliaHashData(override val m: Byte, val r: Long, val s: Long) : HashDataRule
